+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -334,7 +334,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KXGBS9P4BZ280PP4NWTTPZHV"
+        "01KXGH5JA6DQGGHZY24EJ5X6ZX"
     )
     r"""The unique identifier of the data source"""