+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KXGBS9P32FDE6GRCF4EH6APS"
+        "01KXGH5JA49VKRAR8M4BASCN1V"
     )
     r"""The unique identifier of the data source"""