+++ packages/orq-rc/RELEASES.md
@@ -9439,3 +9439,13 @@ Based on:
 - [python v4.12.0-rc.22] packages/orq-rc
 ### Releases
 - [PyPI v4.12.0-rc.22] https://pypi.org/project/orq-ai-sdk/4.12.0-rc.22 - packages/orq-rc
+
+## 2026-07-14 14:43:33
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.790.1 (2.918.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.12.0-rc.23] packages/orq-rc
+### Releases
+- [PyPI v4.12.0-rc.23] https://pypi.org/project/orq-ai-sdk/4.12.0-rc.23 - packages/orq-rc